+++ src/master/metadata_dumper.cc
@@ -24,9 +24,10 @@
 
 #include <string>
 
-#include "common/massert.h"
-#include "common/metadata.h"
-#include "master/filesystem.h"
+#include <common/massert.h>
+#include <master/filesystem.h>
+#include <master/metadata_backend_common.h>
+#include <master/metadata_backend_interface.h>
 
 static bool createPipe(int pipefds[2]) {
 	if (pipe(pipefds) != 0) {
